chore: move items into folders for better organization
