+++ GlucoseTray/Read/GlucoseReader.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.Options;
+﻿using GlucoseTray.Read.Dexcom;
+using GlucoseTray.Read.Nightscout;
+using Microsoft.Extensions.Options;
 
 namespace GlucoseTray.Read;
