+++ GlucoseTray/Read/GlucoseReadingMapper.cs
@@ -1,4 +1,6 @@
 ﻿
+using GlucoseTray.Read.Dexcom;
+using GlucoseTray.Read.Nightscout;
 using Microsoft.Extensions.Options;
 
 namespace GlucoseTray.Read;
