+++ GlucoseTray/Read/IReadStrategy.cs
@@ -3,6 +3,5 @@ namespace GlucoseTray.Read;
 
 public interface IReadStrategy
 {
-
     Task<GlucoseReading> GetLatestGlucoseAsync();
 }
